+++ docs/changelog.rst
@@ -15,6 +15,10 @@ develop
 * [BREAKING-EXPERIMENTAL] The ``batch_identifiers`` key on ``DataConnectorQuery`` (formerly ``PartitionRequest``) has been changed to ``batch_filter_parameters``
 * [ENHANCEMENT] Added a new ``RuntimeBatchRequest`` class, which can be used alongside ``RuntimeDataConnector`` to specify batches at runtime with either an in-memory dataframe, path (filesystem or s3), or sql query
 * [ENHANCEMENT] Added a new ``RuntimeQueryBatchSpec`` class
+* [ENHANCEMENT] CLI `docs list` command implemented for v3 api
+* [ENHANCEMENT] CLI `docs build` command implemented for v3 api
+* [ENHANCEMENT] CLI `docs clean` command implemented for v3 api
+* [MAINTENANCE] Add testing for overwrite_existing in sanitize_yaml_and_save_datasource #2613
 * [BUGFIX] Fixed issue where Sorters were not being applied correctly when ``data_connector_query`` contained limit or index  #2617
 
 0.13.15